doubles e fixes

+++ source/classes/Beacons.h
@@ -10,8 +10,8 @@ using namespace std;
 /* Farol */
 struct Beacon{
     Point position;
-    int radius;
-    int height;
+    double radius;
+    double height;
 };
 
 class Beacons {
