+++ source/classes/Grid.h
@@ -5,7 +5,7 @@
 using namespace std;
 
 struct Pose{
-    int x, y, teta; /* teta => orientação do robot */
+    double x, y, teta; /* teta => orientação do robot por isso tem de estar entre -180 e 180 */
 };
 
 class Grid {
@@ -13,7 +13,7 @@ class Grid {
     vector<Pose>  *posesList;
 
 	public:
-    Grid(){ posesList = new vector<Pose>();}
+    Grid(){ posesList = new vector<Pose>(); }
     void addPose(Pose add);
     void removePose(Pose removePose);
     void modifyPose(Pose toModify, Pose newPose);
